Create Parse-with-lynx-BSO-Jail-Inmate-Information-System.txt
This is a recipe to parse dynamically created web pages.

The example used in this recipe is the BIBB County Jail Inmate Information System. This recipe saves each entry (who was jailed, when?) in a single text file with the corresponding ID and the date.

Please read their site policy before parsing their website programmatically.